backend/requirements.txt
# Base dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-docx==1.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic-settings==2.1.0
minio==7.2.0
pinecone-client==2.2.4
redis==5.0.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
langchain==0.0.350
openai==1.3.5
anthropic==0.7.4

# AI and ML dependencies
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.2.1
transformers==4.36.2
numpy==1.24.3
scikit-learn==1.3.2
sentence-transformers==2.2.2 